corrected issue with animation rate slider

+++ PEP-APP/PEP-APP/Primary VC/GameViewController.h
@@ -14,7 +14,7 @@
 -(void)setBTTreshold:(float)value;
 -(void)setBTBoost:(float)value;
 -(void)setRepetitionCount:(int)value;
--(void)setBreathLength:(int)value;
+-(void)setBreathLength:(float)value;
 -(void)setImageSoundEffect:(NSString*)value;
 -(void)test:(float)value;
 @end
